Relative pointer protocol

+++ src/backend/winit/input.rs
@@ -50,14 +50,14 @@ impl Device for WinitVirtualDevice {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`KeyboardKeyEvent`]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitKeyboardInputEvent {
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) key: u32,
     pub(crate) count: u32,
     pub(crate) state: ElementState,
 }
 
 impl Event<WinitInput> for WinitKeyboardInputEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -84,12 +84,12 @@ impl KeyboardKeyEvent<WinitInput> for WinitKeyboardInputEvent {
 #[derive(Debug, Clone)]
 pub struct WinitMouseMovedEvent {
     pub(crate) size: Rc<RefCell<WindowSize>>,
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) logical_position: LogicalPosition<f64>,
 }
 
 impl Event<WinitInput> for WinitMouseMovedEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -127,12 +127,12 @@ impl AbsolutePositionEvent<WinitInput> for WinitMouseMovedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`PointerAxisEvent`]
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub struct WinitMouseWheelEvent {
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) delta: MouseScrollDelta,
 }
 
 impl Event<WinitInput> for WinitMouseWheelEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -169,14 +169,14 @@ impl PointerAxisEvent<WinitInput> for WinitMouseWheelEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`PointerButtonEvent`]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitMouseInputEvent {
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) button: WinitMouseButton,
     pub(crate) state: ElementState,
     pub(crate) is_x11: bool,
 }
 
 impl Event<WinitInput> for WinitMouseInputEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -210,13 +210,13 @@ impl PointerButtonEvent<WinitInput> for WinitMouseInputEvent {
 #[derive(Debug, Clone)]
 pub struct WinitTouchStartedEvent {
     pub(crate) size: Rc<RefCell<WindowSize>>,
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) location: LogicalPosition<f64>,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchStartedEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -261,13 +261,13 @@ impl AbsolutePositionEvent<WinitInput> for WinitTouchStartedEvent {
 #[derive(Debug, Clone)]
 pub struct WinitTouchMovedEvent {
     pub(crate) size: Rc<RefCell<WindowSize>>,
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) location: LogicalPosition<f64>,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchMovedEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -311,12 +311,12 @@ impl AbsolutePositionEvent<WinitInput> for WinitTouchMovedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a `TouchUpEvent`
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitTouchEndedEvent {
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchEndedEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
 
@@ -336,12 +336,12 @@ impl TouchEvent<WinitInput> for WinitTouchEndedEvent {
 /// Winit-Backend internal event wrapping `winit`'s types into a [`TouchCancelEvent`]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct WinitTouchCancelledEvent {
-    pub(crate) time: u32,
+    pub(crate) time: u64,
     pub(crate) id: u64,
 }
 
 impl Event<WinitInput> for WinitTouchCancelledEvent {
-    fn time(&self) -> u32 {
+    fn time(&self) -> u64 {
         self.time
     }
